Move code generation into askama_derive
